Fix Corp AI trying to play a Hedge Fund when under 5 credits
Corp AI didn't check if the cost is payable.
Historically, we'd be able to directly ask if `hedgeFund.cost.CanPay(game)`. It wasn't enough for UI, because we cannot keep asking this every frame.
So we added an observer pattern, which would notify an observer whenever "payability" of the cost changes.
We had two methods for inspecting payability and the UI only needed the observer-based one and the direct one wasn't used anywhere, so we removed it.
For the Corp AI, we could reintroduce the direct way, but for now I kept the observer-first architecture and introduced "action" usability tracking on the AI level.
So now the Corp will trigger random legal actions. Currently it only supports gaining credits, playing events and installing in existing remotes.