Refactor card-printing components to share state
Cards are printed in many places. All go through `CardPrinter`.
So it can wire in the click-listener to a *shared* zoom-in view.
But it was a `MonoBehavior`, added (and so constructed) in many places.
Injecting a shared view would have to be done in every call site.
The compiler wouldn't help catch missed spots, because `Monobehavior`s don't use constructors.

So I refactored many components into plain old objects, with constructor injection.
I also split the `CardPrinter` initialization into stages.
It starts with just the shared state.
Then it can be attached to a local state (gameObject).

Add a rudimentary `CardZoom` GUI proof of concept.
Currently it has the following problems:
- it cannot be closed
- it will reveal facedown cards
- it creates gameobjects, but it never destroys them

Progress on #104.